refactor: extract datasource forms into standalone components
Break the monolithic AddSourceModal (974 lines) into focused, self-contained
form components following the existing DbtSourceForm/GithubFileSourceForm pattern:

- UploadSourceForm: CSV/XLSX file drag-drop upload
- BigQuerySourceForm: credentials, project/dataset/table selection
- GoogleSheetsSourceForm: spreadsheet ID and sheet selection
- SqlSourceForm: connection string, table discovery and selection

AddSourceModal is now a thin orchestrator (~110 lines) that composes tabs
and delegates connect actions via imperative refs.
